C - Use unsigned and long where sensible
Updated C code to allow long file length of a parsed JSON file.
Added a define to allow long string length if you want that also.

Used unsigned instead of int anywhere a negative doesn't make sense.

+++ c/test.c
@@ -82,7 +82,7 @@ int main( int argc, char *argv[] ) {
         exit(0);
     }
     if( !strncmp(cmd,"test",4) ) {
-        int len;
+        unsigned long len;
         char *data = slurp_file( "test.json", &len );
         int err;
         node_hash *root = parse( data, len, NULL, &err );
